Move tree logic to afumigatus module
