+++ simulation/cell.py
@@ -1,29 +1,23 @@
-from collections.abc import MutableMapping
-from typing import Any, Iterable, Optional, Type, Union
+from typing import Any, Iterable, Type, Union
 
 import attr
+from h5py import Group
 import numpy as np
-from scipy.sparse import dok_matrix as sparse_matrix
-from scipy.spatial.transform import Rotation
 
 from simulation.coordinates import Point
-from simulation.state import RectangularGrid
+from simulation.state import RectangularGrid, State
 
-MAX_CELL_TREE_SIZE = 10000
+MAX_CELL_LIST_SIZE = 10000
 
 # the way numpy types single records is strange...
 CellType = Any
 
 
-class CellArray(np.ndarray):
+class CellData(np.ndarray):
     GROWTH_SCALE_FACTOR = 0.02  # from original code
 
     BASE_FIELDS = [
         ('point', Point.dtype),
-        ('growth', Point.dtype),
-        ('growable', 'b1'),
-        ('switched', 'b1'),
-        ('branchable', 'b1'),
     ]
 
     # typing for dtype doesn't work correctly with this argument
@@ -43,25 +37,8 @@ class CellArray(np.ndarray):
     def create_cell(cls, point: Point = None, **kwargs) -> np.record:
         if point is None:
             point = Point()
-        growth = cls.GROWTH_SCALE_FACTOR * Point.from_array(2 * np.random.rand(3) - 1)
-        growable = True
-        switched = False
-        branchable = False
 
-        return np.rec.array([(point, growth, growable, switched, branchable)], dtype=cls.dtype)[0]
-
-    @classmethod
-    def random_branch_direction(cls, growth: np.ndarray) -> np.ndarray:
-        """Rotate by a random 45 degree angle from the axis of growth."""
-        growth = Point.from_array(growth)
-
-        # get a random vector orthogonal to the growth vector
-        axis = Point.from_array(np.cross(growth, np.random.randn(3)))
-        axis = (np.pi / 4) * axis / axis.norm()
-
-        # rotate the growth vector 45 degrees along the random axis
-        rotation = Rotation.from_rotvec(axis)
-        return rotation.apply(growth)
+        return np.rec.array([(point,)], dtype=cls.dtype)[0]
 
     @classmethod
     def point_mask(cls, points: np.ndarray, grid: RectangularGrid):
@@ -84,197 +61,79 @@ class CellArray(np.ndarray):
         )
 
 
-@attr.s(auto_attribs=True, kw_only=True, frozen=True)
-class Cell(MutableMapping):
-    tree: 'CellTree'
-    index: int = attr.ib()
-
-    @index.validator
-    def _validate_index(self, attribute, value):
-        if value >= len(self.tree):
-            raise ValueError('Invalid cell index')
-
-    def __attrs_post_init__(self):
-        if self.index < 0:
-            object.__setattr__(self, 'index', self.index + len(self.tree))
-
-    @property
-    def record(self) -> CellType:
-        return self.tree.cells[self.index]
-
-    @property
-    def parent(self) -> Optional['Cell']:
-        indices = self.tree.adjacency[:, self.index].nonzero()[0]
-        if len(indices) == 1:
-            return None
-        if len(indices) == 2 and indices[0] == self.index:
-            index = indices[0]
-            if index == self.index:
-                index = indices[1]
-            return self.__class__(tree=self.tree, index=index)
-
-        raise Exception(f'Invalid adjacency matrix at index {self.index}')
-
-    @property
-    def children(self) -> Iterable['Cell']:
-        indices = self.tree.adjacency[self.index].nonzero()[1]
-        for index in indices:
-            if index != self.index:
-                yield self.__class__(tree=self.tree, index=index)
-
-    def add_child(self, cell: Union[CellType, 'Cell']):
-        if isinstance(cell, Cell):
-            cell = cell.record
-        self.tree.append(cell, parent=self.index)
-
-    def __getitem__(self, key):
-        return self.record.__getitem__(key)
-
-    def __setitem__(self, key, value):
-        self.tree.cells[key][self.index] = value
-
-    def __delitem__(self, key):
-        return self.record.__delitem__(key)
-
-    def __iter__(self):
-        for key in self.record.dtype.fields.keys():
-            yield key
-
-    def __contains__(self, key):
-        return key in self.record.dtype.fields
-
-    def __len__(self):
-        return len(self.record)
-
-
 @attr.s(kw_only=True, frozen=True, repr=False)
-class CellTree(object):
-    CellArrayClass: Type[CellArray] = CellArray
+class CellList(object):
+    CellDataClass: Type[CellData] = CellData
 
-    max_cells: int = attr.ib(default=MAX_CELL_TREE_SIZE)
+    max_cells: int = attr.ib(default=MAX_CELL_LIST_SIZE)
     grid: RectangularGrid = attr.ib()
-    _cells: CellArray = attr.ib()
-    _adjacency: sparse_matrix = attr.ib()
+    _cell_data: CellData = attr.ib()
     _ncells: int = attr.ib(init=False)
 
-    @_cells.default
-    def __set_default_cells(self) -> CellArray:
-        return self.CellArrayClass(0)
-
-    @_adjacency.default
-    def __set_default_adjacency(self):
-        return sparse_matrix((0, 0))
+    @_cell_data.default
+    def __set_default_cells(self) -> CellData:
+        return self.CellDataClass(0)
 
     def __attrs_post_init__(self):
-        cells = self._cells
-        adjacency = self._adjacency
+        cells = self._cell_data
 
         object.__setattr__(self, '_ncells', len(cells))
-        object.__setattr__(self, '_cells', self.CellArrayClass(self.max_cells))
-        object.__setattr__(self, '_adjacency', sparse_matrix((self.max_cells, self.max_cells)))
+        object.__setattr__(self, '_cell_data', self.CellDataClass(self.max_cells))
 
         if len(cells) > 0:
-            self._cells[: len(cells)] = cells
-            for key, value in adjacency.items():
-                self._adjacency[key] = value
+            self._cell_data[: len(cells)] = cells
 
     def __len__(self) -> int:
         return self._ncells
 
     def __repr__(self) -> str:
-        return f'CellTree[{self._ncells}]'
+        return f'CellList[{self._ncells}]'
 
-    def __getitem__(self, index: int) -> Cell:
+    def __getitem__(self, index: int) -> CellType:
         if isinstance(index, str):
-            raise TypeError('Expected an integer index, did you mean `tree.cells[key]`?')
-        return Cell(tree=self, index=index)
+            raise TypeError('Expected an integer index, did you mean `cells.cell_data[key]`?')
+        return self.cell_data[index]
 
     @property
-    def root(self) -> Optional[CellType]:
-        if not len(self):
-            return None
-        return self[0]
-
-    @property
-    def cells(self) -> CellArray:
-        return self._cells[: self._ncells]
-
-    @property
-    def adjacency(self) -> sparse_matrix:
-        return self._adjacency[: self._ncells, : self._ncells]
+    def cell_data(self) -> CellData:
+        return self._cell_data[: self._ncells]
 
     @classmethod
-    def create_from_seed(cls, grid, **kwargs) -> 'CellTree':
-        cell = cls.CellArrayClass.create_cell(**kwargs)
-        cells = cls.CellArrayClass([cell])
-        adjacency = sparse_matrix((1, 1))
-        adjacency[0, 0] = 1
+    def create_from_seed(cls, grid, **kwargs) -> 'CellList':
+        cell = cls.CellDataClass.create_cell(**kwargs)
+        cell_data = cls.CellDataClass([cell])
 
-        return cls(grid=grid, cells=cells, adjacency=adjacency)
+        return cls(grid=grid, cell_data=cell_data)
 
-    def append(self, cell: CellType, parent: int = None) -> None:
+    def append(self, cell: CellType) -> None:
         if len(self) >= self.max_cells - 1:
             raise Exception('Not enough free space in cell tree')
 
         index = self._ncells
         object.__setattr__(self, '_ncells', self._ncells + 1)
-        self._cells[index] = cell
-        self._adjacency[index, index] = 1
+        self._cell_data[index] = cell
 
-        if parent is not None:
-            self._adjacency[parent, index] = 1
+    def extend(self, cells: CellData) -> None:
+        for cell in cells:
+            self.append(cell)
 
-    def extend(self, cells: CellArray, parents: Iterable[Union[int, None]] = None) -> None:
-        if parents is None:
-            parents = [None] * len(cells)
+    def save(self, group: Group, name: str, metadata: dict) -> Group:
+        composite_group = group.create_group(name)
 
-        for cell, parent in zip(cells, parents):
-            self.append(cell, parent)
+        class_ = self.__class__
+        composite_group.attrs['type'] = 'CellList'
+        composite_group.attrs['class'] = f'{class_.__module__}:{class_.__name__}'
+        composite_group.attrs['max_cells'] = self.max_cells
 
-    def is_growable(self) -> np.ndarray:
-        cells = self.cells
-        return cells['growable'] & cells.point_mask(cells['point'] + cells['growth'], self.grid)
+        composite_group.create_dataset(name='cell_data', data=self.cell_data)
+        return composite_group
 
-    def is_branchable(self, branch_probability: float) -> np.ndarray:
-        cells = self.cells
-        return cells['branchable'] & (np.random.rand(*cells.shape) < branch_probability)
+    @classmethod
+    def load(cls, global_state: State, group: Group, name: str, metadata: dict) -> 'CellList':
+        composite_dataset = group[name]
 
-    def elongate(self):
-        cells = self.cells
-        mask = self.is_growable().nonzero()[0]
-        if len(mask) == 0:
-            return
+        attrs = composite_dataset.attrs
+        max_cells = attrs.get('max_cells', MAX_CELL_LIST_SIZE)
+        cell_data = composite_dataset['cell_data'][:].view(cls.CellDataClass)
 
-        cells['growable'][mask] = False
-        cells['branchable'][mask] = True
-
-        children = self.CellArrayClass(len(mask), initialize=True)
-        children['point'] = cells['point'][mask] + cells['growth'][mask]
-        children['growth'] = cells['growth'][mask]
-
-        self.extend(children, parents=mask)
-
-    def branch(self, branch_probability: float):
-        cells = self.cells
-        indices = self.is_branchable(branch_probability).nonzero()[0]
-        if len(indices) == 0:
-            return
-
-        children = self.CellArrayClass(len(indices), initialize=True)
-        children['growth'] = np.apply_along_axis(
-            cells.random_branch_direction, 1, children['growth']
-        )
-
-        children['point'] = cells['point'][indices] + children['growth']
-
-        # filter out children lying outside of the domain
-        indices = indices[cells.point_mask(children['point'], self.grid)]
-        if len(indices) == 0:
-            return
-
-        # set properties
-        cells['branchable'][indices] = False
-        children['growable'] = True
-        children['branchable'] = False
-
-        self.extend(children, parents=indices)
+        return cls(max_cells=max_cells, grid=global_state.grid, cell_data=cell_data)
